# Rework BaseMaterial3D deep parallax implementation and add new features