The deep parallax effect now features built-in interpolation to make low layer counts look better. It also integrates Z correction for a more convincing depth effect at oblique angles.

The new features are disabled by default, but allow for greater realism at a performance cost. Each of them can be enabled separately:

- **Correct Shadow Receive:** This allows the material to receive shadows (and even cast self-shadows) according to the deep parallax effect. Self-shadowing is powered by the existing shadow map system, so it works best with sharp shadows such as the ones cast by OmniLight3D and SpotLight3D, especially when placed at grazing angles. This only has a small performance cost, but it can cause self-shadowing artifacts if lights have a shadow bias that's too low or if the material has a heightmap scale that's too high. Enabling **Write Depth** at the same time can alleviate those self-shadowing artifacts, with the associated performance cost of that feature.
  - For a self-shadowing implementation that works directly in the shader (without even needing lights to have shadows enabled), https://github.com/godotengine/godot-proposals/issues/8671 would have to be implemented first.

- **Write Depth:** This allows the material to correctly interact with other materials in the scene, regardless of whether the other materials have deep parallax enabled. This has a significant performance cost.

- **Trim Edges:** This allows the material to have its silhouette affected by the deep parallax effect. This works best with certain UV layouts, but it can have artifacts if the UV layout has visible seams. This has a moderate performance cost.

Thanks to @sphynx who made the [initial implementation](https://github.com/sphynx-owner/parallax-mapping) and Tentabrobpy from the Godot VFX Discord for further assistance :slightly_smiling_face:

- This partially addresses https://github.com/godotengine/godot-proposals/issues/1843.

**Testing project:** https://github.com/Calinou/godot-parallax-test-4.0

## Preview

![Parallax 1](https://github.com/user-attachments/assets/3c1c11f4-a85c-4968-9f6f-51ce97976b27) | ![Parallax 2](https://github.com/user-attachments/assets/fd6489c8-4bb3-4eb6-8943-1f642ae96191)
-|-
![Parallax 3](https://github.com/user-attachments/assets/99a46690-b8b0-4cb1-8db6-a465ad44d59a) | ![Parallax 4](https://github.com/user-attachments/assets/04140729-7985-44ad-b870-aea4a4517e12)

https://github.com/user-attachments/assets/dd05ffc9-a40e-4caa-8562-c9df72c06be0

https://github.com/user-attachments/assets/5571e5e3-002d-436d-86f9-e77c8ae1c291

### Comparison with default settings

### 8/32 layers (default quality)

Before | After
-|-
![Before](https://github.com/user-attachments/assets/228df17b-e802-4021-bcf5-221169ca93f3) | ![After](https://github.com/user-attachments/assets/e85d34c5-03fb-4875-8adc-13c61f1d4458)

### 64/64 layers (maximum quality)

Before | After
-|-
![Before](https://github.com/user-attachments/assets/30ec7021-d8b4-446f-9ca8-c7a888ee5370) | ![After](https://github.com/user-attachments/assets/6d01558d-95a1-4361-9e31-43c9e27856ed)

### Visual impact of each feature being toggled

None | CSR[^1] | WD[^2] | 
-|-|-
![none png webp](https://github.com/user-attachments/assets/42730844-5656-4cda-90a7-9575357c4c67) | ![csr png webp](https://github.com/user-attachments/assets/041d5ffe-498c-4c56-810e-3d13f9fd1c57) | ![wd png webp](https://github.com/user-attachments/assets/f4642b84-209a-4613-a57f-f2524a571e77)

TE[^3] | CSR[^1] + WD[^2] | CSR[^1] + TE[^3] |
-|-|-
![te png webp](https://github.com/user-attachments/assets/159d3777-0b3f-495a-8a58-568357c8c991) | ![csr_wd png webp](https://github.com/user-attachments/assets/a11ae028-2f1f-4b89-8474-f31e4ee48462) | ![csr_te png webp](https://github.com/user-attachments/assets/34afa493-b20c-4002-bbbf-9f5cd0d91af5)

WD[^2] + TE[^3] | CSR[^1] + WD[^2] + TE[^3]
-|-
![wd_te png webp](https://github.com/user-attachments/assets/ccd596f2-fa73-4ced-b379-4480ddea3a75) | ![csr_wd_te png webp](https://github.com/user-attachments/assets/32d6cf8a-4a9a-4ea3-b100-881d1c301f23)

[^1]: Correct Shadow Receive
[^2]: Write Depth
[^3]: Trim Edges

## Benchmark

<details open>

<summary>PC specifications</summary>

- **CPU:** Intel Core i9-13900K
- **GPU:** NVIDIA GeForce RTX 4090
- **RAM:** 64 GB (2×32 GB DDR5-5800 C30)
- **SSD:** Solidigm P44 Pro 2 TB
- **OS:** Linux (Fedora 40)
</details>

*Measured using the [test scene](https://github.com/Calinou/godot-parallax-test-4.0
) shown in the Before/After comparison before. Materials cover most (but not all) of the view. Using a 3840×2160 viewport.*

Features | Performance
-|-
None | 843 FPS (1.19 mspf)
CSR[^1] | 856 FPS (1.17 mspf)
WD[^2] | 351 FPS (2.85 mspf)
TE[^3] | 689 FPS (1.45 mspf)
CSR[^1] + WD[^2] | 350 FPS (2.86 mspf)
CSR[^1] + TE[^3] | 689 FPS (1.45 mspf)
WD[^2] + TE[^3] | 341 FPS (2.93 mspf)
CSR[^1] + WD[^2] + TE[^3] | 344 FPS (2.91 mspf)

## TODO

- [ ] When **Write Depth** is enabled, see if we can disable the depth prepass without having to use `discard` within `if (false)`. Since **Write Depth** does not actually require using `discard`, doing so would probably improve performance a bit.
  - [ ] Or, better, see if the depth prepass can be kept when **Write Depth** is enabled, by having the depth prepass shrink the material's vertices (only within the depth prepass), so that it can still be effective and improve performance significantly.
- [ ] Check the interactions with the UV Scale property again to ensure **Correct Shadow Receive** looks as good as possible regardless of the scale. Right now, it can get a bit off when the scale is changed from the default (i.e. when the texture tiles more often), although it's not too noticeable during actual gameplay.
- [ ] Consider combining **Correct Shadow Receive** and **Write Depth** into a single enum option, since there isn't much point in having **Write Depth** enabled but not **Correct Shadow Receive**. This means there would be 3 options provided: **Disabled**, **Correct Shadow Receive**, **Correct Shadow Receive + Write Depth**.