@Calinou 2024-12-10
I've pushed an update that makes use of @tetrapod00's suggested fix for Compatibility. However, it causes a shader compilation error:

```
SHADER ERROR: Tokenizer: Unknown character #35: '#'
          at: (null) (:96)
ERROR: Shader compilation failed.
   at: set_code (./drivers/gles3/storage/material_storage.cpp:2983)
```

For some reason, preprocessor macros don't work in the shader generated in BaseMaterial3D, even though it doesn't use `#pragma disable_preprocessor`. They start working if you convert it to a ShaderMaterial.