@solitaryurt 2025-02-05
MacOS x86_64, looks like its not working:
Deep Parallax disabled:
<img width="754" alt="image" src="https://github.com/user-attachments/assets/1ae02654-4a9e-4df9-9629-be355074f6cd" />

Deep Parallax enabled:
<img width="754" alt="image" src="https://github.com/user-attachments/assets/a6d4fae7-3c19-46d3-897e-1501f8f4aa82" />

Used build from: https://github.com/Calinou/godot/commit/1f6e871795d5b775862746086f0b9437252d08fd