@solitaryurt 2025-02-05
> > MacOS x86_64 (Forward+), looks like its not working:
> 
> This is expected for now until we fix BaseMaterial3D to allow using shader preprocessor within its generated code. This should be done in a separate PR.

Is there an older hash I could build without the shader compilation issue?