@Calinou 2025-02-05
> Is there an older hash I could build without the shader compilation issue?

No, but you can remove the uses of the shader preprocessor within the PR's changes if you only intend to use Forward+/Mobile in your project.