@solitaryurt 2025-02-05
None:
![image](https://github.com/user-attachments/assets/e89934f8-d19e-4825-8803-1e5c4a0626d0)

Correct Shadows:
![image](https://github.com/user-attachments/assets/70db44dc-5ec1-4c5f-b6f5-1c1d21500b4f)

Write Depth:
![image](https://github.com/user-attachments/assets/127e232c-1c9c-40e3-8872-f91dc8fdb673)

Correct Shadows + Write Depth:
![image](https://github.com/user-attachments/assets/e1a35145-8d4f-479b-9052-edd3913b8b93)

Unfortunately the mesh I'm using isn't happy with Trim Edges:
![image](https://github.com/user-attachments/assets/c2b9dc10-a98a-475a-a769-1827db9bf825)