@solitaryurt 2025-02-05
I'm sure Trim Edges and Write Depth will meet the needs of many people who desired displacement mapping.

![image](https://github.com/user-attachments/assets/1960317c-e83b-44f5-a499-982693265858)
![image](https://github.com/user-attachments/assets/0aa500bd-b73b-4dd6-bc46-c0b899162f45)

![image](https://github.com/user-attachments/assets/e95ff857-a47a-40f9-a958-af505607a03e)
![image](https://github.com/user-attachments/assets/97c73b99-a076-44c6-b3b1-20e0f6105cb6)

I will note that I believe the addition of these features makes https://github.com/godotengine/godot-proposals/issues/2455 even more relevant now.