@Calinou 2025-02-06
> Unfortunately the mesh I'm using isn't happy with Trim Edges:

It doesn't work well with repeating textures unfortunately. I'm not aware of a way to make it work with repeating textures (unless the number of texture repetitions is known in advance).