@solitaryurt 2025-02-09
I noticed that when "Deep Parallax" is enabled both "Flip Tangent" and "Flip Binormal" don't appear to have an effect. Is that intended?