@solitaryurt 2025-02-09
Is there any possibility of anti-aliasing the edges Trim Edges or Write Depth?

I noticed some aliasing:
![image](https://github.com/user-attachments/assets/8d0f61df-a040-4792-8969-238644ca6c8d)