@Calinou 2025-02-10
> Is there any possibility of anti-aliasing the edges Trim Edges or Write Depth?

Not by MSAA, because the edges are created by a shader and not by geometry. However, all other antialiasing methods (TAA, FXAA, FSR2, SSAA, [SMAA](https://github.com/godotengine/godot/pull/102330)) will work.